pyproject.toml
[project]
name = "scholarsnap"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "certifi>=2025.10.5",
    "dotenv>=0.9.9",
    "feedparser>=6.0.12",
    "flower>=2.0.1",
    "google-api-python-client>=2.186.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "langchain>=1.0.3",
    "langchain-google-community[gmail]>=3.0.0",
    "langchain-openai>=1.0.1",
    "psycopg2>=2.9.11",
    "pypdf>=6.1.3",
    "python-dotenv>=1.2.1",
    "redis>=7.0.1",
    "requests>=2.32.5",
    "schedule>=1.2.2",
]
